Initial database syncing
The source in the JavaServerSocket folder sets up a listening server on
the computer. It needs a path to the location of the master db on the
computer.

Change the port numbers and host addresses for your network setup, and
run the ggdatabase server, and connect by pressing the Generate QR code
button in New Record.